# CDRIVER-4740 bson_utf8_escape_for_json performance improvements