--- a/src/libbson/src/bson/bson-string.c
+++ b/src/libbson/src/bson/bson-string.c
@@ -126,15 +126,11 @@ bson_string_new (const char *str) /* IN */
  *
  *       Create an empty bson_string_t and allocate memory for it.
  *
- *       bson_string_t is a power-of-2 allocation growing string. Every
- *       time data is appended the next power of two size is chosen for
- *       the allocation. Pretty standard stuff.
- *
- *       It is UTF-8 aware through the use of bson_string_append_unichar().
- *       The proper UTF-8 character sequence will be used.
+ *       The amount of memory allocated will be the next power-of-two if the
+ *       specified size is not already a power-of-two.
  *
  * Parameters:
- *       @len: Length of the string to allocate
+ *       @size: Size of the string to allocate
  *
  * Returns:
  *       A newly allocated bson_string_t that should be freed with
@@ -147,16 +143,19 @@ bson_string_new (const char *str) /* IN */
  */
 
 bson_string_t *
-bson_string_alloc (uint8_t len)
+bson_string_alloc (const uint32_t size)
 {
    bson_string_t *ret;
 
    ret = bson_malloc0 (sizeof *ret);
    ret->len = 0;
-   ret->alloc = len + 1;
+   ret->alloc = size + 1;
 
    if (!bson_is_power_of_two (ret->alloc)) {
-      ret->alloc = (uint32_t) bson_next_power_of_two ((size_t) ret->alloc);
+      ret->alloc = bson_next_power_of_two_u32 (ret->alloc);
+      if (ret->alloc == 0) {
+         ret->alloc = __UINT32_MAX__;
+      }
    }
 
    BSON_ASSERT (ret->alloc >= 1);