--- a/src/libbson/src/bson/bson-string.h
+++ b/src/libbson/src/bson/bson-string.h
@@ -39,8 +39,12 @@ typedef struct {
 
 BSON_EXPORT (bson_string_t *)
 bson_string_new (const char *str);
+BSON_EXPORT (bson_string_t *)
+bson_string_alloc (uint8_t len);
 BSON_EXPORT (char *)
 bson_string_free (bson_string_t *string, bool free_segment);
+BSON_EXPORT (void)
+bson_string_append_ex (bson_string_t *string, const char *str, size_t len);
 BSON_EXPORT (void)
 bson_string_append (bson_string_t *string, const char *str);
 BSON_EXPORT (void)
@@ -49,6 +53,8 @@ BSON_EXPORT (void)
 bson_string_append_unichar (bson_string_t *string, bson_unichar_t unichar);
 BSON_EXPORT (void)
 bson_string_append_printf (bson_string_t *string, const char *format, ...) BSON_GNUC_PRINTF (2, 3);
+BSON_EXPORT (void)
+bson_string_append_codepoint (bson_string_t *string, bson_unichar_t unichar);
 BSON_EXPORT (void)
 bson_string_truncate (bson_string_t *string, uint32_t len);
 BSON_EXPORT (char *)