--- a/src/libbson/src/bson/bson-string.h
+++ b/src/libbson/src/bson/bson-string.h
@@ -54,8 +54,6 @@ bson_string_append_unichar (bson_string_t *string, bson_unichar_t unichar);
 BSON_EXPORT (void)
 bson_string_append_printf (bson_string_t *string, const char *format, ...) BSON_GNUC_PRINTF (2, 3);
 BSON_EXPORT (void)
-bson_string_append_codepoint (bson_string_t *string, bson_unichar_t unichar);
-BSON_EXPORT (void)
 bson_string_truncate (bson_string_t *string, uint32_t len);
 BSON_EXPORT (char *)
 bson_strdup (const char *str);