--- a/src/libbson/src/bson/bson-string.h
+++ b/src/libbson/src/bson/bson-string.h
@@ -40,11 +40,11 @@ typedef struct {
 BSON_EXPORT (bson_string_t *)
 bson_string_new (const char *str);
 BSON_EXPORT (bson_string_t *)
-bson_string_alloc (uint32_t len);
+bson_string_alloc (const uint32_t len);
 BSON_EXPORT (char *)
 bson_string_free (bson_string_t *string, bool free_segment);
 BSON_EXPORT (void)
-bson_string_append_ex (bson_string_t *string, const char *str, size_t len);
+bson_string_append_ex (bson_string_t *string, const char *str, const size_t len);
 BSON_EXPORT (void)
 bson_string_append (bson_string_t *string, const char *str);
 BSON_EXPORT (void)