--- a/src/libbson/src/bson/bson.c
+++ b/src/libbson/src/bson/bson.c
@@ -1472,7 +1472,7 @@ bson_append_regex_w_len (
       options = "";
    }
 
-   options_sorted = bson_string_new (NULL);
+   options_sorted = bson_string_alloc (strlen (options));
 
    _bson_append_regex_options_sorted (options_sorted, options);