… src/libbson/doc/bson_string_alloc.rst
… src/libbson/doc/bson_string_append_codepoint.rst
… src/libbson/doc/bson_string_t.rst