# Summary
This PR optimizes `bson_utf8_escape_for_json` to reduce performance problems in `bson_as_json`.

# Motivation
Credit to @alcaeus for the initial find and [draft PR](https://github.com/mongodb/mongo-c-driver/pull/1419). Much of the work in this PR is adapted from his changes. To summarize, `bson_as_json` was performing about 5x as slow when compared to the PHP driver's equivalent function. It was recognized that `bson_utf8_escape_for_json` was the main reason for this.

# Background
In order to convert a UTF-8 string to JSON, some characters require [special treatment](https://www.json.org/json-en.html). For example, double quotes `"` must be converted into escaped double quotes `\"`. See [this diagram](https://www.json.org/img/string.png) for all of the necessary conversions. As such, when converting to JSON, strings must be parsed character by character to ensure that the resulting string is valid for JSON.

# `_is_special_char()`
Previously, `bson_utf8_escape_for_json` was checking, converting, and copying over each character one-by-one. Considering that most characters do not require any conversion, a bit mask is now used to see if a character is "special" or not. If not, we simply continue to iterate. When a special character is reached (or end of string), we append of ALL of the normal characters seen up to that point. This reduces the amount of appends to the string, improving performance. `bson_string_append_ex` was added as a helper for this piece-by-piece strategy.

# `bson_string_alloc()`
Previously, an empty string was created in `bson_utf8_escape_for_json` and then appended to in order to produce the JSON-safe string. Being that we can guarantee that the string produced from this function will be at least the size of the string passed in, this is wasteful due to the cost of growing the string from size 0. To fix this, a helper function `bson_string_alloc` was created. This reduces the amount of reallocations needed as we now create a initial string with the minimal known size.

# Testing
Unit tests and docs were added for the new public API `bson_string_alloc` and  `bson_string_append_ex`. In addition to that and running the `utf8` tests, the fuzz tester for `bson_utf8_escape_for_json` successfully ran without issues, which was the initial reason for backlogging the [draft PR](https://github.com/mongodb/mongo-c-driver/pull/1419). This difference in behavior is likely due to the fix from [this](https://github.com/mongodb/mongo-c-driver/pull/1725) PR.

# Performance Improvements
To test performance, [these](https://github.com/kevinAlbs/c-bootstrap/tree/master/investigations/profileme) profiling steps were followed using both a `bson_utf8_escape_for_json` [benchmark](https://gist.github.com/joshbsiegel/cd783aed9e80481c8a2fa2b7766399eb) and a `bson_as_json` [benchmark](https://gist.github.com/joshbsiegel/61b2847b34cc22cee20510dde20816af).

| Benchmark    | Time Before (s) |  Time After (s) | Improvement | 
| :------------: | :-------: | :-------: | :-------: |
| `bson_utf8_escape_for_json`  | 35.94  ![bson_utf8_escape_for_json before](https://github.com/user-attachments/assets/be3fd344-7d15-428c-b13f-57dbc143fe39)  | 13.16 ![bson_utf8_escape_for_json after](https://github.com/user-attachments/assets/c6911fce-8f48-4ff3-8d09-0a953ed71137) | 63.38%
| `bson_as_json` | 56.04 ![bson_as_json_before](https://github.com/user-attachments/assets/fa48612c-8257-4193-bf20-628ec2339cbc)     | 35.54 ![bson_as_json_after](https://github.com/user-attachments/assets/27a44c26-eb0a-43ac-a621-f6afb5d8f07b) | 36.58%

# What's New
- `bson_string_alloc()` implementation, documentation, and tests
- `bson_string_append_ex()` implementation, documentation, and tests
- `bson_utf8_escape_for_json()` optimizations (see above) and refactor